Added fertilityGrid and made timeM 1 = 1 hour.
TimeM 1 = 1 h. Its now set to minutes.
The grid is 10x10 meters. The human walks 5 km/h. Growthspeed of weeds is 1/1000th of what it should be right now. Movementspeed of tumbleWeeds is incorrect. and should be fixed.

+++ Assets/Scripts/TimeMultipier.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public class TimeMultipier : MonoBehaviour
 {
-    public int timeMultipier;
+    public float timeMultipier;
 }
